code dedup in rendergraph and RenderGraphPanel
