+++ webgpu/base/gpu_utils.cpp
@@ -167,4 +167,316 @@ WGPUBindGroupEntry bind(uint32_t binding, WGPUBuffer buffer, uint64_t offset, ui
     return e;
 }
 
+// ---- WebGPU texture-format math ------------------------------------------------------------------
+
+// block size in texels (1x1 for uncompressed) plus bytes per block
+TexelBlock format_block(WGPUTextureFormat f)
+{
+    switch (f)
+    {
+    // 8-bit
+    case WGPUTextureFormat_R8Unorm:
+    case WGPUTextureFormat_R8Snorm:
+    case WGPUTextureFormat_R8Uint:
+    case WGPUTextureFormat_R8Sint:
+    case WGPUTextureFormat_Stencil8:
+        return { 1, 1, 1 };
+
+    // 16-bit
+    case WGPUTextureFormat_R16Uint:
+    case WGPUTextureFormat_R16Sint:
+    case WGPUTextureFormat_R16Float:
+
+    case WGPUTextureFormat_RG8Unorm:
+    case WGPUTextureFormat_RG8Snorm:
+    case WGPUTextureFormat_RG8Uint:
+    case WGPUTextureFormat_RG8Sint:
+
+    case WGPUTextureFormat_Depth16Unorm:
+        return { 1, 1, 2 };
+
+    // 32-bit
+    case WGPUTextureFormat_R32Float:
+    case WGPUTextureFormat_R32Uint:
+    case WGPUTextureFormat_R32Sint:
+
+    case WGPUTextureFormat_RG16Uint:
+    case WGPUTextureFormat_RG16Sint:
+    case WGPUTextureFormat_RG16Float:
+
+    case WGPUTextureFormat_RGBA8Unorm:
+    case WGPUTextureFormat_RGBA8UnormSrgb:
+    case WGPUTextureFormat_RGBA8Snorm:
+    case WGPUTextureFormat_RGBA8Uint:
+    case WGPUTextureFormat_RGBA8Sint:
+
+    case WGPUTextureFormat_BGRA8Unorm:
+    case WGPUTextureFormat_BGRA8UnormSrgb:
+
+    case WGPUTextureFormat_RGB10A2Uint:
+    case WGPUTextureFormat_RGB10A2Unorm:
+    case WGPUTextureFormat_RG11B10Ufloat:
+
+    case WGPUTextureFormat_Depth24Plus:
+    case WGPUTextureFormat_Depth32Float:
+        return { 1, 1, 4 };
+
+    // 64-bit
+    case WGPUTextureFormat_RG32Float:
+    case WGPUTextureFormat_RG32Uint:
+    case WGPUTextureFormat_RG32Sint:
+
+    case WGPUTextureFormat_RGBA16Uint:
+    case WGPUTextureFormat_RGBA16Sint:
+    case WGPUTextureFormat_RGBA16Float:
+
+    case WGPUTextureFormat_Depth24PlusStencil8:
+    case WGPUTextureFormat_Depth32FloatStencil8:
+        return { 1, 1, 8 };
+
+    // 128-bit
+    case WGPUTextureFormat_RGBA32Float:
+    case WGPUTextureFormat_RGBA32Uint:
+    case WGPUTextureFormat_RGBA32Sint:
+        return { 1, 1, 16 };
+
+
+    // BC1 / BC4: 8 bytes per 4x4 block
+    case WGPUTextureFormat_BC1RGBAUnorm:
+    case WGPUTextureFormat_BC1RGBAUnormSrgb:
+    case WGPUTextureFormat_BC4RUnorm:
+    case WGPUTextureFormat_BC4RSnorm:
+        return { 4, 4, 8 };
+
+    // BC2/3/5/6H/7: 16 bytes per 4x4 block
+    case WGPUTextureFormat_BC2RGBAUnorm:
+    case WGPUTextureFormat_BC2RGBAUnormSrgb:
+    case WGPUTextureFormat_BC3RGBAUnorm:
+    case WGPUTextureFormat_BC3RGBAUnormSrgb:
+    case WGPUTextureFormat_BC5RGUnorm:
+    case WGPUTextureFormat_BC5RGSnorm:
+    case WGPUTextureFormat_BC6HRGBUfloat:
+    case WGPUTextureFormat_BC6HRGBFloat:
+    case WGPUTextureFormat_BC7RGBAUnorm:
+    case WGPUTextureFormat_BC7RGBAUnormSrgb:
+        return { 4, 4, 16 };
+
+
+    // ETC2 RGB8 / RGB8A1 and EAC R11: 8 bytes per 4x4 block
+    case WGPUTextureFormat_ETC2RGB8Unorm:
+    case WGPUTextureFormat_ETC2RGB8UnormSrgb:
+    case WGPUTextureFormat_ETC2RGB8A1Unorm:
+    case WGPUTextureFormat_ETC2RGB8A1UnormSrgb:
+    case WGPUTextureFormat_EACR11Unorm:
+    case WGPUTextureFormat_EACR11Snorm:
+        return { 4, 4, 8 };
+
+    // ETC2 RGBA8 and EAC RG11: 16 bytes per 4x4 block
+    case WGPUTextureFormat_ETC2RGBA8Unorm:
+    case WGPUTextureFormat_ETC2RGBA8UnormSrgb:
+    case WGPUTextureFormat_EACRG11Unorm:
+    case WGPUTextureFormat_EACRG11Snorm:
+        return { 4, 4, 16 };
+
+
+    // ASTC: always 16 bytes per block, block footprint read straight off the enum name
+    case WGPUTextureFormat_ASTC4x4Unorm:
+    case WGPUTextureFormat_ASTC4x4UnormSrgb:   return { 4, 4, 16 };
+    case WGPUTextureFormat_ASTC5x4Unorm:
+    case WGPUTextureFormat_ASTC5x4UnormSrgb:   return { 5, 4, 16 };
+    case WGPUTextureFormat_ASTC5x5Unorm:
+    case WGPUTextureFormat_ASTC5x5UnormSrgb:   return { 5, 5, 16 };
+    case WGPUTextureFormat_ASTC6x5Unorm:
+    case WGPUTextureFormat_ASTC6x5UnormSrgb:   return { 6, 5, 16 };
+    case WGPUTextureFormat_ASTC6x6Unorm:
+    case WGPUTextureFormat_ASTC6x6UnormSrgb:   return { 6, 6, 16 };
+    case WGPUTextureFormat_ASTC8x5Unorm:
+    case WGPUTextureFormat_ASTC8x5UnormSrgb:   return { 8, 5, 16 };
+    case WGPUTextureFormat_ASTC8x6Unorm:
+    case WGPUTextureFormat_ASTC8x6UnormSrgb:   return { 8, 6, 16 };
+    case WGPUTextureFormat_ASTC8x8Unorm:
+    case WGPUTextureFormat_ASTC8x8UnormSrgb:   return { 8, 8, 16 };
+    case WGPUTextureFormat_ASTC10x5Unorm:
+    case WGPUTextureFormat_ASTC10x5UnormSrgb:  return { 10, 5, 16 };
+    case WGPUTextureFormat_ASTC10x6Unorm:
+    case WGPUTextureFormat_ASTC10x6UnormSrgb:  return { 10, 6, 16 };
+    case WGPUTextureFormat_ASTC10x8Unorm:
+    case WGPUTextureFormat_ASTC10x8UnormSrgb:  return { 10, 8, 16 };
+    case WGPUTextureFormat_ASTC10x10Unorm:
+    case WGPUTextureFormat_ASTC10x10UnormSrgb: return { 10, 10, 16 };
+    case WGPUTextureFormat_ASTC12x10Unorm:
+    case WGPUTextureFormat_ASTC12x10UnormSrgb: return { 12, 10, 16 };
+    case WGPUTextureFormat_ASTC12x12Unorm:
+    case WGPUTextureFormat_ASTC12x12UnormSrgb: return { 12, 12, 16 };
+
+    default:
+        break;
+    }
+
+    return { 0, 0, 0 };
+}
+
+uint64_t texture_bytes(WGPUExtent3D size, WGPUTextureFormat format, uint32_t mipLevelCount, uint32_t sampleCount, WGPUTextureDimension dim)
+{
+    const TexelBlock b = format_block(format);
+    if (!b.bytes) return 0;
+    const bool is3D = dim == WGPUTextureDimension_3D;
+    const uint32_t layers = is3D ? 1u : (size.depthOrArrayLayers ? size.depthOrArrayLayers : 1);
+    const uint32_t samples = sampleCount ? sampleCount : 1;
+    uint64_t total = 0;
+    for (uint32_t m = 0; m < mipLevelCount; ++m) {
+        const uint32_t w = (size.width  >> m) ? (size.width  >> m) : 1u;
+        const uint32_t h = (size.height >> m) ? (size.height >> m) : 1u;
+        const uint32_t d = is3D ? ((size.depthOrArrayLayers >> m) ? (size.depthOrArrayLayers >> m) : 1u) : 1u;
+        const uint32_t bw = (w + b.w - 1) / b.w;
+        const uint32_t bh = (h + b.h - 1) / b.h;
+        total += (uint64_t)bw * bh * d * layers * samples * b.bytes;
+    }
+    return total;
+}
+
+const char* format_name(WGPUTextureFormat f)
+{
+    switch (f)
+    {
+    // 8-bit formats
+    case WGPUTextureFormat_R8Unorm:              return "R8Unorm";
+    case WGPUTextureFormat_R8Snorm:              return "R8Snorm";
+    case WGPUTextureFormat_R8Uint:               return "R8Uint";
+    case WGPUTextureFormat_R8Sint:               return "R8Sint";
+
+    // 16-bit formats
+    case WGPUTextureFormat_R16Uint:              return "R16Uint";
+    case WGPUTextureFormat_R16Sint:              return "R16Sint";
+    case WGPUTextureFormat_R16Float:             return "R16Float";
+
+    case WGPUTextureFormat_RG8Unorm:             return "RG8Unorm";
+    case WGPUTextureFormat_RG8Snorm:             return "RG8Snorm";
+    case WGPUTextureFormat_RG8Uint:              return "RG8Uint";
+    case WGPUTextureFormat_RG8Sint:              return "RG8Sint";
+
+    // 32-bit formats
+    case WGPUTextureFormat_R32Float:             return "R32Float";
+    case WGPUTextureFormat_R32Uint:              return "R32Uint";
+    case WGPUTextureFormat_R32Sint:              return "R32Sint";
+
+    case WGPUTextureFormat_RG16Uint:             return "RG16Uint";
+    case WGPUTextureFormat_RG16Sint:             return "RG16Sint";
+    case WGPUTextureFormat_RG16Float:            return "RG16Float";
+
+    case WGPUTextureFormat_RGBA8Unorm:           return "RGBA8Unorm";
+    case WGPUTextureFormat_RGBA8UnormSrgb:       return "RGBA8UnormSrgb";
+    case WGPUTextureFormat_RGBA8Snorm:           return "RGBA8Snorm";
+    case WGPUTextureFormat_RGBA8Uint:            return "RGBA8Uint";
+    case WGPUTextureFormat_RGBA8Sint:            return "RGBA8Sint";
+
+    case WGPUTextureFormat_BGRA8Unorm:           return "BGRA8Unorm";
+    case WGPUTextureFormat_BGRA8UnormSrgb:       return "BGRA8UnormSrgb";
+
+    case WGPUTextureFormat_RGB10A2Uint:          return "RGB10A2Uint";
+    case WGPUTextureFormat_RGB10A2Unorm:         return "RGB10A2Unorm";
+    case WGPUTextureFormat_RG11B10Ufloat:        return "RG11B10Ufloat";
+
+    // 64-bit formats
+    case WGPUTextureFormat_RG32Float:            return "RG32Float";
+    case WGPUTextureFormat_RG32Uint:             return "RG32Uint";
+    case WGPUTextureFormat_RG32Sint:             return "RG32Sint";
+
+    case WGPUTextureFormat_RGBA16Uint:           return "RGBA16Uint";
+    case WGPUTextureFormat_RGBA16Sint:           return "RGBA16Sint";
+    case WGPUTextureFormat_RGBA16Float:          return "RGBA16Float";
+
+    // 128-bit formats
+    case WGPUTextureFormat_RGBA32Float:          return "RGBA32Float";
+    case WGPUTextureFormat_RGBA32Uint:           return "RGBA32Uint";
+    case WGPUTextureFormat_RGBA32Sint:           return "RGBA32Sint";
+
+    // depth / stencil
+    case WGPUTextureFormat_Stencil8:             return "Stencil8";
+    case WGPUTextureFormat_Depth16Unorm:         return "Depth16Unorm";
+    case WGPUTextureFormat_Depth24Plus:          return "Depth24Plus";
+    case WGPUTextureFormat_Depth24PlusStencil8: return "Depth24PlusStencil8";
+    case WGPUTextureFormat_Depth32Float:         return "Depth32Float";
+    case WGPUTextureFormat_Depth32FloatStencil8:return "Depth32FloatStencil8";
+
+    // BC compressed
+    case WGPUTextureFormat_BC1RGBAUnorm:         return "BC1RGBAUnorm";
+    case WGPUTextureFormat_BC1RGBAUnormSrgb:     return "BC1RGBAUnormSrgb";
+    case WGPUTextureFormat_BC2RGBAUnorm:         return "BC2RGBAUnorm";
+    case WGPUTextureFormat_BC2RGBAUnormSrgb:     return "BC2RGBAUnormSrgb";
+    case WGPUTextureFormat_BC3RGBAUnorm:         return "BC3RGBAUnorm";
+    case WGPUTextureFormat_BC3RGBAUnormSrgb:     return "BC3RGBAUnormSrgb";
+    case WGPUTextureFormat_BC4RUnorm:            return "BC4RUnorm";
+    case WGPUTextureFormat_BC4RSnorm:            return "BC4RSnorm";
+    case WGPUTextureFormat_BC5RGUnorm:           return "BC5RGUnorm";
+    case WGPUTextureFormat_BC5RGSnorm:           return "BC5RGSnorm";
+    case WGPUTextureFormat_BC6HRGBUfloat:        return "BC6HRGBUfloat";
+    case WGPUTextureFormat_BC6HRGBFloat:         return "BC6HRGBFloat";
+    case WGPUTextureFormat_BC7RGBAUnorm:         return "BC7RGBAUnorm";
+    case WGPUTextureFormat_BC7RGBAUnormSrgb:     return "BC7RGBAUnormSrgb";
+
+    // ETC2
+    case WGPUTextureFormat_ETC2RGB8Unorm:             return "ETC2RGB8Unorm";
+    case WGPUTextureFormat_ETC2RGB8UnormSrgb:         return "ETC2RGB8UnormSrgb";
+    case WGPUTextureFormat_ETC2RGB8A1Unorm:           return "ETC2RGB8A1Unorm";
+    case WGPUTextureFormat_ETC2RGB8A1UnormSrgb:       return "ETC2RGB8A1UnormSrgb";
+    case WGPUTextureFormat_ETC2RGBA8Unorm:            return "ETC2RGBA8Unorm";
+    case WGPUTextureFormat_ETC2RGBA8UnormSrgb:        return "ETC2RGBA8UnormSrgb";
+    case WGPUTextureFormat_EACR11Unorm:               return "EACR11Unorm";
+    case WGPUTextureFormat_EACR11Snorm:               return "EACR11Snorm";
+    case WGPUTextureFormat_EACRG11Unorm:              return "EACRG11Unorm";
+    case WGPUTextureFormat_EACRG11Snorm:              return "EACRG11Snorm";
+
+    // ASTC
+    case WGPUTextureFormat_ASTC4x4Unorm:              return "ASTC4x4Unorm";
+    case WGPUTextureFormat_ASTC4x4UnormSrgb:          return "ASTC4x4UnormSrgb";
+    case WGPUTextureFormat_ASTC5x4Unorm:              return "ASTC5x4Unorm";
+    case WGPUTextureFormat_ASTC5x4UnormSrgb:          return "ASTC5x4UnormSrgb";
+    case WGPUTextureFormat_ASTC5x5Unorm:              return "ASTC5x5Unorm";
+    case WGPUTextureFormat_ASTC5x5UnormSrgb:          return "ASTC5x5UnormSrgb";
+    case WGPUTextureFormat_ASTC6x5Unorm:              return "ASTC6x5Unorm";
+    case WGPUTextureFormat_ASTC6x5UnormSrgb:          return "ASTC6x5UnormSrgb";
+    case WGPUTextureFormat_ASTC6x6Unorm:              return "ASTC6x6Unorm";
+    case WGPUTextureFormat_ASTC6x6UnormSrgb:          return "ASTC6x6UnormSrgb";
+    case WGPUTextureFormat_ASTC8x5Unorm:              return "ASTC8x5Unorm";
+    case WGPUTextureFormat_ASTC8x5UnormSrgb:          return "ASTC8x5UnormSrgb";
+    case WGPUTextureFormat_ASTC8x6Unorm:              return "ASTC8x6Unorm";
+    case WGPUTextureFormat_ASTC8x6UnormSrgb:          return "ASTC8x6UnormSrgb";
+    case WGPUTextureFormat_ASTC8x8Unorm:              return "ASTC8x8Unorm";
+    case WGPUTextureFormat_ASTC8x8UnormSrgb:          return "ASTC8x8UnormSrgb";
+    case WGPUTextureFormat_ASTC10x5Unorm:             return "ASTC10x5Unorm";
+    case WGPUTextureFormat_ASTC10x5UnormSrgb:         return "ASTC10x5UnormSrgb";
+    case WGPUTextureFormat_ASTC10x6Unorm:             return "ASTC10x6Unorm";
+    case WGPUTextureFormat_ASTC10x6UnormSrgb:         return "ASTC10x6UnormSrgb";
+    case WGPUTextureFormat_ASTC10x8Unorm:             return "ASTC10x8Unorm";
+    case WGPUTextureFormat_ASTC10x8UnormSrgb:         return "ASTC10x8UnormSrgb";
+    case WGPUTextureFormat_ASTC10x10Unorm:            return "ASTC10x10Unorm";
+    case WGPUTextureFormat_ASTC10x10UnormSrgb:        return "ASTC10x10UnormSrgb";
+    case WGPUTextureFormat_ASTC12x10Unorm:            return "ASTC12x10Unorm";
+    case WGPUTextureFormat_ASTC12x10UnormSrgb:        return "ASTC12x10UnormSrgb";
+    case WGPUTextureFormat_ASTC12x12Unorm:            return "ASTC12x12Unorm";
+    case WGPUTextureFormat_ASTC12x12UnormSrgb:        return "ASTC12x12UnormSrgb";
+
+    default:
+        break;
+    }
+
+    return ""; 
+}
+
+
+const char* format_name_short(WGPUTextureFormat f)
+{
+    switch (f) {
+      case WGPUTextureFormat_RGBA8Unorm:   return "RGBA8";
+      case WGPUTextureFormat_BGRA8Unorm:   return "BGRA8";
+      case WGPUTextureFormat_RGBA16Float:  return "RGBA16F";
+      case WGPUTextureFormat_R32Float:     return "R32F";
+      case WGPUTextureFormat_Depth32Float: return "D32F";
+      case WGPUTextureFormat_Depth24Plus:  return "D24+";
+      case WGPUTextureFormat_R8Unorm:      return "R8";
+      default:                             return "tex";
+    }
+}
+
 } // namespace webgpu
