+++ webgpu/base/gpu_utils.h
@@ -19,6 +19,7 @@
 
 #pragma once
 
+#include <cstdint>
 #include <webgpu/webgpu.h>
 
 namespace webgpu {
@@ -37,5 +38,31 @@ void compute_mipmaps_for_texture(Context& ctx, const raii::Texture* texture, WGP
 WGPUBindGroupEntry bind(uint32_t binding, WGPUTextureView view);
 WGPUBindGroupEntry bind(uint32_t binding, WGPUSampler sampler);
 WGPUBindGroupEntry bind(uint32_t binding, WGPUBuffer buffer, uint64_t offset, uint64_t size);
+
+
+// block size in texels (1x1 for uncompressed) plus bytes per block
+struct TexelBlock {
+    uint32_t w, h, bytes;
+};
+
+// texel-block descriptor for a format. { 0, 0, 0 } for an unknown format
+TexelBlock format_block(WGPUTextureFormat f);
+
+// exact size in bytes of one texture, summed over its mip chain
+uint64_t texture_bytes(WGPUExtent3D size, WGPUTextureFormat format, uint32_t mipLevelCount = 1, uint32_t sampleCount = 1, WGPUTextureDimension dim = WGPUTextureDimension_2D);
+
+// the 256-byte row alignment WebGPU requires for buffer<->texture copies
+inline uint32_t aligned_bytes_per_row(uint32_t widthTexels, uint32_t texelBlockBytes)
+{
+    constexpr uint32_t align = 256u;
+    uint32_t raw = widthTexels * texelBlockBytes;
+    return (raw + (align - 1u)) & ~(align - 1u);
+}
+
+// full format label (e.g. "RGBA8Unorm"), unknown formats return an empty string
+const char* format_name(WGPUTextureFormat f);
+
+// abbreviated format label for compact UI (e.g. "RGBA8"), unknown formats fall back to "tex"
+const char* format_name_short(WGPUTextureFormat f);
 
 } // namespace webgpu
